% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{celeb}
\alias{celeb}
\title{Celeb Names}
\format{
\subsection{\code{celeb}}{

A data frame with 4557 rows and 5 columns:
\describe{
\item{name}{name}
\item{catagory}{occupation}
\item{image}{url link to an image of this celebrity}
\item{well_known_movie}{What movie the celeb is known for}
\item{bio}{Description of the life of the celebrity}

...
}
}
}
\source{
\url{https://www.kaggle.com/code/rishabhdhyani4/celebrity/data}
}
\usage{
celeb
}
\description{
Data that contains 4557 celebrity names, occupations, roles etc.
}
\keyword{datasets}
