% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celeb_name.R
\name{celeb_name}
\alias{celeb_name}
\title{celeb_name: Function that finds all the celebrities in a particular movie.}
\usage{
celeb_name(movie)
}
\arguments{
\item{movie}{a character}
}
\value{
character vector
}
\description{
celeb_name: Function that finds all the celebrities in a particular movie.
}
\examples{
celeb_name("Slumdog Millionaire")
}
