% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_year_numeric.R
\name{check_year_numeric}
\alias{check_year_numeric}
\title{Check Year Numeric}
\usage{
check_year_numeric(y)
}
\arguments{
\item{y}{user input, second argument}
}
\description{
Check Year Numeric
}
\examples{
check_year_numeric(2000000)
}
