% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_year_range.R
\name{check_year_range}
\alias{check_year_range}
\title{Check Year Range}
\usage{
check_year_range(y)
}
\arguments{
\item{y}{Numeric}
}
\value{
Nothing.
}
\description{
Check Year Range
}
\examples{
check_year_range(1999)
}
