% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spanish_female}
\alias{spanish_female}
\title{Spanish Female Names}
\format{
\subsection{\code{spanish_female}}{

A data frame with 24756 rows and 3 columns:
\describe{
\item{name}{name}
\item{frequency}{frequency}
\item{mean_age}{mean age}
...
}
}
}
\source{
\url{http://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736177009&menu=resultados&secc=1254736195454&idp=1254734710990}

\url{https://github.com/jvalhondo/spanish-names-surnames}
}
\usage{
spanish_female
}
\description{
Data that contains 24746 Spanish first names given to current
female Spanish residents and newborns birthed in the past
15 years. Cleaned by jvalhondo on github.
}
\keyword{datasets}
