% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppercase_input.R
\name{uppercase_input}
\alias{uppercase_input}
\title{uppercase input: changes character input to uppercase}
\usage{
uppercase_input(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character (uppercase)
}
\description{
uppercase input: changes character input to uppercase
}
